requirements.txt
PyYAML==6.0
python-dateutil==2.8.2
google-cloud-storage==2.8.0
google-cloud-bigquery==3.9.0
google-api-core==2.11.0